images/django/requirements.txt
Django==4.2.7
djangorestframework
Pillow
psycopg2
pathlib
django.urls
bcrypt
requests
python-dotenv
